refactor: roll back to single header single implementation
This commit refactor the whole lib to roll to a single header
implementation, using a single implementation and ditching the
interface. This is to make the code simpler and more efficient as well
as simpler.
