+++ README.md
@@ -2,9 +2,7 @@
 Modern C++ Thread Pool
 
 # Available features
-* Multiples Threadpool implementation:
-  * Single queue multiple workers Threadpool.
-  * Multiple queue multiple workers Threadpool.
+* Multiple queue multiple workers Threadpool.
 * Fixed/Variable size pool (Adapt number of workers when under heavy load).
 * Fetch results of tasks run in the pool with `future`.
 * Fetch number of waiting/working workers.
@@ -14,8 +12,7 @@ Modern C++ Thread Pool
 
 ## Github Releases
 
-You can get the latest pre-compiled releases for Linux x86\_64 and Mac OSX from
-the github release page.
+You can get the latest releases for Linux x86\_64 and Mac OSX from the github release page.
 
 The latest release can be found [here](https://github.com/reyreaud-l/threadpool/releases/latest).
 
@@ -49,26 +46,10 @@ You will find multiples examples and a how to get started.
 ## Compatibility
 ThreadPool is tested on these platforms/compilers:
 - Linux
-  - gcc
-    - 4.9
-    - 5
-    - 6
-    - 7
-  - clang
-    - 3.6
-    - 3.7
-    - 3.8
-    - 3.9
-    - 4.0
-    - 5.0
+  - gcc: 4.9, 5, 6, 7, 8
+  - clang: 3.6, 3.7, 3.8, 3.9, 4.0, 5.0, 6.0
 - OSX
-  - gcc
-    - 5
-    - 6
-    - 7
-  - clang xcode
-    - 8.3
-    - 9.1
+  - clang xcode: 6.4, 7.3, 8, 8.1, 8.2, 8.3, 9, 9.1, 9.2, 9.3, 9.4, 10
 
 # Contributing ThreadPool
 
@@ -88,4 +69,15 @@ $ cmake -DBUILD_TESTS=ON ..
 Then the `check` target becomes available and you can build it running
 ```bash
 $ make check
+```
+
+## Examples
+To build the examples, enable option `BUILD_EXAMPLES`:
+```bash
+$ cmake -DBUILD_EXAMPLES=ON ..
+```
+
+Then the `examples` target becomes available and you can build it running
+```bash
+$ make examples
 ```
