+++ bench/CMakeLists.txt
@@ -6,10 +6,12 @@ add_executable(benchbin
 )
 
 target_link_libraries(benchbin PRIVATE
-  threadpool
   Threads::Threads
   benchmark
 )
-target_include_directories(benchbin PRIVATE ${CMAKE_SOURCE_DIR}/bench)
+target_include_directories(benchbin PRIVATE
+  ${CMAKE_SOURCE_DIR}/bench
+  ${CMAKE_SOURCE_DIR}/include
+)
 
 add_custom_target(bench benchbin DEPENDS benchbin COMMENT "Running benchs...")
