+++ bench/main.cpp
@@ -1,13 +1,14 @@
 #include "benchmark/benchmark.h"
 #include "threadpool.hpp"
+
 #include <set>
 #include <vector>
 
-static void BM_ThreadPool_SingleQueue_NoBlockingTasks(benchmark::State& state)
+static void BM_ThreadPool_NoBlockingTasks(benchmark::State& state)
 {
   while (state.KeepRunning())
   {
-    ThreadPool::ThreadPool<ThreadPool::SingleQueue> pool;
+    ThreadPool::ThreadPool pool;
     std::vector<std::future<int>> results;
     std::set<std::thread::id> threads_id;
 
@@ -19,28 +20,8 @@ static void BM_ThreadPool_SingleQueue_NoBlockingTasks(benchmark::State& state)
   }
 }
 
-BENCHMARK(BM_ThreadPool_SingleQueue_NoBlockingTasks)
-  ->Range(8, 8 << 14)
-  ->Unit(benchmark::kMillisecond);
-
-static void BM_ThreadPool_MultipleQueue_NoBlockingTasks(benchmark::State& state)
-{
-  while (state.KeepRunning())
-  {
-    ThreadPool::ThreadPool<ThreadPool::MultipleQueue> pool;
-    std::vector<std::future<int>> results;
-    std::set<std::thread::id> threads_id;
-
-    auto i_end = state.range(0);
-    for (int i = 0; i < i_end; i++)
-      results.push_back(pool.run([i]() { return i * i; }));
-    for (int i = 0; i < i_end; i++)
-      results[i].get();
-  }
-}
-
-BENCHMARK(BM_ThreadPool_MultipleQueue_NoBlockingTasks)
-  ->Range(8, 8 << 14)
+BENCHMARK(BM_ThreadPool_NoBlockingTasks)
+  ->Range(8, 8 << 20)
   ->Unit(benchmark::kMillisecond);
 
 BENCHMARK_MAIN();
