+++ cmake/ClangTidy.cmake
@@ -1,9 +1,5 @@
 macro(add_tidy_sources)
-  get_target_property(_SOURCES ${ARGV0} SOURCES)
-  set(TIDY_SOURCES "")
-  foreach(SRC ${_SOURCES})
-    get_filename_component(ABS_SRC ${SRC} ABSOLUTE)
-    list(APPEND TIDY_SOURCES "${ABS_SRC}")
-  endforeach()
+  get_filename_component(ABS_SRC ${ARGV0} ABSOLUTE)
+  list(APPEND TIDY_SOURCES "${ABS_SRC}")
   set(TIDY_SOURCES "${TIDY_SOURCES}" PARENT_SCOPE)
 endmacro()
