+++ tests/CMakeLists.txt
@@ -3,16 +3,17 @@ include(GTest)
 add_executable(tests
   EXCLUDE_FROM_ALL
   main.cpp
-  single_queue.cpp
   multiple_queue.cpp
   hooks.cpp
 )
 
 target_link_libraries(tests PRIVATE
-  threadpool
   gtest_main
   Threads::Threads
 )
-target_include_directories(tests PRIVATE ${CMAKE_SOURCE_DIR}/tests)
+target_include_directories(tests PRIVATE
+  ${CMAKE_SOURCE_DIR}/tests
+  ${CMAKE_SOURCE_DIR}/include
+)
 
 add_custom_target(check tests DEPENDS tests COMMENT "Running tests...")
